documentation con't

+++ ConsoleHero/ParagraphBuilder.cs
@@ -5,9 +5,8 @@
 public static class ParagraphBuilder
 {
     /// <summary>
-    /// 
+    /// Clear the console when this <see cref="INode"/> is called.
     /// </summary>
-    /// <returns></returns>
     public static ISetLines ClearOnCall() => new Builder().ClearOnCall();
 
     /// <summary>
@@ -62,8 +61,7 @@ public static class ParagraphBuilder
         /// 
         /// </summary>
         /// <param name="seconds"></param>
-        /// <returns></returns>
-        public Paragraph DelayInSeconds(int seconds);
+        public Paragraph DelayInSeconds(double seconds);
         /// <summary>
         /// After displaying this paragraph, will wait for the user to press a key to continue.
         /// </summary>
@@ -98,7 +96,7 @@ public static class ParagraphBuilder
         /// </summary>
         /// <param name="seconds"></param>
         /// <returns></returns>
-        public Paragraph DelayInSeconds(int seconds);
+        public Paragraph DelayInSeconds(double seconds);
         /// <summary>
         /// After displaying this paragraph, will wait for the user to press a key to continue.
         /// </summary>
@@ -144,7 +142,7 @@ public static class ParagraphBuilder
             _item.PressToContinue = false;
             return _item;
         }
-        public Paragraph DelayInSeconds(int seconds)
+        public Paragraph DelayInSeconds(double seconds)
         {
             _item.Delay = TimeSpan.FromSeconds(seconds);
             _item.PressToContinue = false;
